requirements.txt
streamlit==1.45.0
pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.12.0
umap-learn>=0.5.3
torch>=2.0.0
torch-geometric>=2.3.0
FlowCal==1.3.1
